validate screenshot dimensions
