+++ package.json
@@ -11,7 +11,8 @@
     "build:pdf": "magicbook build",
     "test": "jest",
     "test:examples": "jest tests/examples.test.js",
-    "test:snippets": "jest tests/snippets.test.js"
+    "test:snippets": "jest tests/snippets.test.js",
+    "test:screenshots": "jest tests/screenshots.test.js"
   },
   "repository": {
     "type": "git",
@@ -58,6 +59,7 @@
     "rehype-parse": "^8.0.4",
     "rehype-react": "^7.1.1",
     "rehype-slug": "^5.0.1",
+    "sharp": "^0.32.6",
     "tailwindcss": "^3.1.6",
     "typescript": "^5.2.2",
     "unified": "^10.1.2",
